requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib
alembic
pydantic
python-dotenv
email-validator
python-multipart